automate collecting configuration from the DMS-10
I failed at speaking the telnet DO/DONT/WILL/WONT protocols, so this just forks
out to `/usr/bin/telnet` for now.

This ended up being a really simplistic `expect`: each OVLY/TYP has a fixed set
of "send this, wait for that prompt", allowing both CR and LF at the beginning
of the line.  Many of the `Fetcher`s are one common pattern, but there ended up
being quite a variety of prompting sequence and exact padding in the DMS-10's
software.

The target filename (e.g. `DN/STN.txt`) is the unique identifier for a fetcher,
and can be specified on the command line to fetch only a subset of the config.

This traps Ctrl-C and allows you to skip, retry, or wait for the current
fetcher to finish.